% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_date_filter.R
\name{generate_date_filter}
\alias{generate_date_filter}
\title{Generate Date Filter}
\usage{
generate_date_filter(start_date, end_date)
}
\arguments{
\item{start_date}{start date for the period of analysis}

\item{end_date}{end date for the period of analysis}
}
\description{
Generate date filter given inputted start/end dates
}
