% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get Socrata API Token}
\usage{
get_token(token_name = "SOCRATA_API_TOKEN")
}
\arguments{
\item{token_name}{The name attached to the Socrata API token in a .Renviron file.}
}
\value{
The Socrata API token stored in a .Renviron file.
}
\description{
Load the Socrata API token stored in the .Renviron file
}
