% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_311.R
\name{import_311}
\alias{import_311}
\title{Import Cincinnati 3-1-1 Calls}
\usage{
import_311(start_date, end_date = Sys.Date(),
  agency_responsible = NULL, api_token)
}
\arguments{
\item{start_date}{start date for period of analysis}

\item{end_date}{end date for period of analysis}

\item{agency_responsible}{character vector consisting of government agencies of interest}

\item{api_token}{Socrata API token used for querying the API}
}
\value{
data frame consisting of 3-1-1 call data based on input filters
}
\description{
Easily import 3-1-1 data with a set of filters
}
\examples{
# create objects to be used as filter arguments
start_date <- '2019-10-01'
end_date <- '2019-10-10'
agency_responsible <- c('Public Services', 'Police Department')
socrata_token <- get_token()

# call function with selected filter arguments
import_311(start_date = start_date,
           end_date = end_date,
           agency_responsible = agency_responsible,
           api_token = socrata_token)

}
